engine/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
alpaca-py==0.34.0
pandas==2.2.3
numpy==1.26.4
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
PyYAML==6.0.2
loguru==0.7.3
httpx==0.28.1
aiohttp==3.11.10
pytz==2024.2
